Исправлено разнесение между модулями, отдельные классы --disabled
